pyproject.toml
[tool.poetry]
name = "PyYATL"
version = "0.1.1"
description = "YAML Templating Language"
authors = ["Dan Hipschman"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/d5h-foss/yatl"
repository = "https://github.com/d5h-foss/yatl"
keywords = ["yaml", "template"]
documentation = "https://yatl.readthedocs.io"

packages = [
    { include = "yatl", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
coverage = {extras = ["toml"], version = "^5.3.1"}
pytest-cov = "^2.10.1"
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.1"
flake8-bugbear = "^20.11.1"
flake8-import-order = "^0.18.1"
safety = "^1.10.3"
mypy = "^0.790"
codecov = "^2.1.11"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["yatl"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
